Add optional switch to enable Django compatibility
This commit adds an optional switch to initialize django before
the worker starts. This enables the usage of objects like django models
in runnel processor functions, which is useful if runnel is being used
with a django project. This is not a full-fledged django integration,
which would include features like autodiscovery and integration with
django settings.